Android Production (Finished) and Suppliers (Pending)

+++ EB_AndroidApp/eb_androidapp/app/src/main/java/com/easybusiness/eb_androidapp/Entities/ProductionBatches.java
@@ -38,7 +38,7 @@ public class ProductionBatches implements Serializable {
     private int ID;
     private int QuantityProduced;
     private int ProductID;
-    private int ProductionDate;
+    private long ProductionDate;
 
 	//-------------------- Constructor --------------------
 
@@ -46,7 +46,7 @@ public class ProductionBatches implements Serializable {
 		int ID, 
 		int QuantityProduced, 
 		int ProductID, 
-		int ProductionDate
+		long ProductionDate
 		) {
         this.ID = ID;
 		this.QuantityProduced = QuantityProduced;
@@ -74,7 +74,7 @@ public class ProductionBatches implements Serializable {
 	/**
      * @return int
      */
-     public int getProductionDate() { return this.ProductionDate; }
+     public long getProductionDate() { return this.ProductionDate; }
 
 
 	//-------------------- Setter Methods --------------------
@@ -92,7 +92,7 @@ public class ProductionBatches implements Serializable {
 	/**
      * @param value int(13)
      */
-     public void setProductionDate(int value) { this.ProductionDate = value; }
+     public void setProductionDate(long value) { this.ProductionDate = value; }
 
 
 	//-------------------- JSON Generation Methods --------------------
